requirements.txt
asgiref
braintree
certifi
chardet
Django
idna
Pillow
pytz
requests
sqlparse
urllib3
django-crispy-forms
django-mathfilters
django-allauth
psycopg2
gunicorn
django-heroku
selenium
boto3
django-storages
django_countries
pycountry
django-rosetta
scipy
nltk
pandas
twilio
humanize
django-anymail
sendgrid
djangorestframework
djangorestframework-csv
xhtml2pdf
django-parler
